+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 * Russian translation (thanks to Stepanyaa*)
 
 ### Changed:
-* Journal got tweaked and also translated to spanish (thanks to Shiro)
+* Journal got tweaked and also translated to spanish (thanks to ShiroPlayZ)
 
 ### Fixed:
 * Some console spam from the Player Data sync packet if the player wasn't found.
+* Check screen not working if the target was unloaded from the command player client.
+* Check screen equipped abilities not being displayed properly.
 * Back button from Equipment Check screen opening the local player equipment.
 * Potential crashes or malfunctioning regarding null abilities being unequipped (for addons).